requirements.txt
python-telegram-bot>=20.7
openai>=1.12.0
dateparser>=1.1.8
APScheduler>=3.10.4
fpdf2>=2.7.8
icalendar>=5.0.12
python-dotenv>=1.0.1
pytz>=2024.1
